# BitsWriter: WriteN bugfix; Write and WriteN zero allocation and optimizations