--- a/binary.go
+++ b/binary.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"math"
 )
 
 // BitsWriter represents an object that can write individual bits into a writer
@@ -13,9 +12,11 @@ import (
 // This is particularly helpful when you want to build a slice of bytes based
 // on individual bits for testing purposes.
 type BitsWriter struct {
-	bo binary.ByteOrder
-	rs []rune
-	w  io.Writer
+	bo       binary.ByteOrder
+	cache    byte
+	cacheLen byte
+	//rs []rune
+	w io.Writer
 }
 
 // BitsWriterOptions represents BitsWriter options
@@ -44,70 +45,96 @@ func NewBitsWriter(o BitsWriterOptions) (w *BitsWriter) {
 //   - []byte: processed as n bytes, n being the length of the input
 //   - bool: processed as one bit
 //   - uint8/uint16/uint32/uint64: processed as n bits, if type is uintn
-func (w *BitsWriter) Write(i interface{}) (err error) {
+func (w *BitsWriter) Write(i interface{}) error {
 	// Transform input into "10010" format
-	var s string
+
 	switch a := i.(type) {
 	case string:
-		s = a
+		for _, r := range a {
+			var err error
+			if r == '1' {
+				err = w.writeBit(1)
+			} else {
+				err = w.writeBit(0)
+			}
+			if err != nil {
+				return err
+			}
+		}
 	case []byte:
 		for _, b := range a {
-			s += fmt.Sprintf("%.8b", b)
+			if err := w.writeFullByte(b); err != nil {
+				return err
+			}
 		}
 	case bool:
 		if a {
-			s = "1"
+			return w.writeBit(1)
 		} else {
-			s = "0"
+			return w.writeBit(0)
 		}
 	case uint8:
-		s = fmt.Sprintf("%.8b", i)
+		return w.writeFullByte(a)
 	case uint16:
-		s = fmt.Sprintf("%.16b", i)
+		return w.writeFullInt(uint64(a), 2)
 	case uint32:
-		s = fmt.Sprintf("%.32b", i)
+		return w.writeFullInt(uint64(a), 4)
 	case uint64:
-		s = fmt.Sprintf("%.64b", i)
+		return w.writeFullInt(a, 8)
 	default:
-		err = errors.New("astikit: invalid type")
-		return
+		return errors.New("astikit: invalid type")
 	}
 
-	// Loop through runes
-	for _, r := range s {
-		// Append rune
-		if w.bo == binary.LittleEndian {
-			w.rs = append([]rune{r}, w.rs...)
-		} else {
-			w.rs = append(w.rs, r)
+	return nil
+}
+
+func (w *BitsWriter) writeFullInt(in uint64, len int) error {
+	if w.bo == binary.BigEndian {
+		for i := len - 1; i >= 0; i-- {
+			err := w.writeFullByte(byte((in >> (i * 8)) & 0xff))
+			if err != nil {
+				return err
+			}
 		}
-
-		// There are enough bits to form a byte
-		if len(w.rs) == 8 {
-			// Get value
-			v := w.val()
-
-			// Remove runes
-			w.rs = w.rs[8:]
-
-			// Write
-			if err = binary.Write(w.w, w.bo, v); err != nil {
-				return
+	} else {
+		for i := 0; i < len; i++ {
+			err := w.writeFullByte(byte((in >> (i * 8)) & 0xff))
+			if err != nil {
+				return err
 			}
 		}
 	}
-	return
+
+	return nil
 }
 
-func (w *BitsWriter) val() (v uint8) {
-	var power float64
-	for idx := len(w.rs) - 1; idx >= 0; idx-- {
-		if w.rs[idx] == '1' {
-			v = v + uint8(math.Pow(2, power))
-		}
-		power++
+func (w *BitsWriter) writeFullByte(b byte) error {
+	if w.cacheLen == 0 {
+		_, err := w.w.Write([]byte{b})
+		return err
 	}
-	return
+
+	toWrite := w.cache | (b >> w.cacheLen)
+	if _, err := w.w.Write([]byte{toWrite}); err != nil {
+		return err
+	}
+
+	w.cache = b << (8 - w.cacheLen)
+	return nil
+}
+
+func (w *BitsWriter) writeBit(bit byte) error {
+	w.cache = w.cache | (bit)<<(7-w.cacheLen)
+	w.cacheLen++
+	if w.cacheLen == 8 {
+		_, err := w.w.Write([]byte{w.cache})
+		if err != nil {
+			return err
+		}
+		w.cacheLen = 0
+		w.cache = 0
+	}
+	return nil
 }
 
 // WriteN writes the input into n bits