--- a/binary_test.go
+++ b/binary_test.go
@@ -94,7 +94,6 @@ func BenchmarkBitsWriter(b *testing.B) {
 		{uint16(5)},
 		{uint32(6)},
 		{uint64(7)},
-		{1},
 	}
 
 	bw := &bytes.Buffer{}