@barbashov 2021-02-01
Could you create a tag as well please? To bump dependencies in go.mod :)